% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lss_county_xwalk}
\alias{lss_county_xwalk}
\title{Maryland Local School System-County Crosswalk}
\format{
A data frame with 25 rows and 7 variables:
\describe{
\item{\code{lss_number}}{Local school system (LSS) number as integer.}
\item{\code{lss_name}}{Local school system (LSS) names.}
\item{\code{countyfp}}{County FIPS code.}
\item{\code{countyns}}{County GNIS code.}
\item{\code{geoid}}{Unique county FIPS code (concatenation of state and county
FIPS codes).}
\item{\code{county_name}}{County name. Equivalent to \code{namelsad} column from
\code{mapmaryland::md_counties} which is a concatenated variable length
geographic area name and legal/statistical area description (LSAD)}
\item{\code{region}}{Region of the state.}
}
}
\usage{
lss_county_xwalk
}
\description{
A data frame that can be used as a crosswalk between Maryland Local School
System (LSS) names and numbers and county-level identifiers from the U.S.
Census Bureau.
}
\keyword{datasets}
