% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marylandedu-package.R
\docType{package}
\name{marylandedu-package}
\alias{marylandedu}
\alias{marylandedu-package}
\title{marylandedu: Maryland State Department of Education Data}
\description{
The marylandedu package provides access to selected datasets from the Maryland State Department of Education (MSDE) that have been standardized to better support year-by-year analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/marylandedu}
  \item \url{https://elipousson.github.io/marylandedu/}
  \item Report bugs at \url{https://github.com/elipousson/marylandedu/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
