% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msde_attendance}
\alias{msde_attendance}
\title{Maryland Public School Attendance (SY 2003-2022)}
\format{
A data frame with 33,376 rows and 18 variables:
\describe{
\item{\code{year}}{School year or academic year, e.g. 2019 for 2019-2020
school year}
\item{\code{school_number}}{School number (0 indicates all schools)}
\item{\code{school_name}}{School name}
\item{\code{school_type}}{School type (e.g. Elementary, Middle, High)}
\item{\code{attend_rate_pct}}{Attendance rate as defined as the average daily
attendance rate of students in the school and grade band, including special
education students}
\item{\code{days_attended_count}}{Total number of students in attendance from the
first day of school through April 13th}
\item{\code{days_member_count}}{Total number of students registered to attend from
the first day of school through April 13th}
\item{\code{fewer_5_pct}}{Percent of students absentee rate is the percentage of
students absent from school for five or fewer days between the first day of
the school year and the last day of the school year}
\item{\code{fewer_5_count}}{Count of students absentee rate is the percentage of
students absent from school for five or fewer days between the first day of
the school year and the last day of the school year}
\item{\code{more_20_pct}}{Percent of students absentee rate is the percentage of
students absent from school for twenty or more days between the first day
of the school year and the last day of the school year}
\item{\code{more_20_count}}{Number of students absentee rate is the percentage of
students absent from school for twenty or more days between the first day
of the school year and the last day of the school year}
\item{\code{more_90_member_count}}{Number of students with an attendance rate of
90 percent or more}
\item{\code{chronic_absentee_count}}{Number of students expected to attend school
for at least 10 days who were absent 10 percent or more of the school days
while enrolled at that school}
\item{\code{chronic_absentee_denom}}{Number of students who are expected to
attend school for at least 10 days in the school year}
\item{\code{chronic_absentee_pct}}{Chronic absenteeism rate as defined as the
percent of students expected to attend school for at least 10 days who were
absent 10 percent or more of the school days while enrolled at that school}
\item{\code{date_created}}{Date record created.}
\item{\code{lss_number}}{Local school system (LSS) number as integer. NA is used
in place of a LSS number for statewide data. Older LEA numbers are combined
with LSS numbers in this dataset.}
\item{\code{lss_name}}{Local school system (LSS) names (typically same as county
name). "State" is used in place of a LSS name for statewide data. Older LEA
names are combined with LSS names in this dataset.}
}
}
\source{
\href{https://reportcard.msde.maryland.gov/Graphs/#/DataDownloads/datadownload}{MSDE Data Downloads}
}
\usage{
msde_attendance
}
\description{
Attendance data from the Maryland State Department of Education (MSDE). #'
The original CSV data files downloaded from MSDE are available in the
\code{data-raw/extdata} folder on the GitHub repo for the package. Any percent variables with a value of 95 percent
should be interpreted as equivalent at or above 95 percent. Any percent
variables with a value of 5 percent should be interpreted as equivalent to at
or below 5 percent. Chronic absenteeism data is not available prior to 2018
when the federal Every Student Succeeds Act (passed in 2015) went into effect
and Maryland began to include the chronic absenteeism rate in the state's
annual academic report cards. Variable definitions are based on the
\href{https://reportcard.msde.maryland.gov/Definitions/Index}{definitions from the MSDE website}.
}
\keyword{datasets}
