% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msde_enrollment}
\alias{msde_enrollment}
\title{Maryland Public School Enrollment (SY 2003-2024)}
\format{
A data frame with 247,740 rows and 10 variables:
\describe{
\item{\code{year}}{School or academic year for enrollment count, e.g.
2019 data is from the start of the 2019-2020 school year.}
\item{\code{school_number}}{School number as integer (0 indicates all
schools)}
\item{\code{school_name}}{School name}
\item{\code{enrolled_count}}{Number of students registered to attend the
school at the start of the year in the grade or grade range as integer.
Typically enrollment count is as of September 30.}
\item{\code{grade}}{Grades from Prekindergarden (PK) to Grade 12}
\item{\code{grade_range}}{all elementary school Grades, all middle school
grades, all high school grades, or all grades (total enrollment). May also
be used as a label for grade.}
\item{\code{race}}{Race/ethnicity. Enrollment by race/ethnicity is only
available across all grades for years since 2020. "All" is used for data
prior to 2020.}
\item{\code{date_created}}{Date record created.}
\item{\code{lss_number}}{Local school system (LSS) number as integer. NA is used
in place of a LSS number for statewide data. Older LEA numbers are combined
with LSS numbers in this dataset.}
\item{\code{lss_name}}{Local school system (LSS) names (typically same as county
name). "State" is used in place of a LSS name for statewide data. Older LEA
names are combined with LSS names in this dataset.}
}
}
\source{
\href{https://reportcard.msde.maryland.gov/Graphs/#/DataDownloads/datadownload}{MSDE Data Downloads}
}
\usage{
msde_enrollment
}
\description{
Enrollment data from the Maryland State Department of Education (MSDE). The
original CSV data files downloaded from MSDE are available in the
\code{data-raw/extdata} folder on the GitHub repo for the package. The number of enrolled students includes ungraded
special education and pre-kindergarten students. Grades coded as
"Prekindergarten Age 4" in the MSDE data have been re-coded as
"Prekindergarten" to facilitate comparison across years. LEA Names coded as
"Baltimore City - Edison" have been re-coded as "Baltimore City", "Seed
School LEA" as "SEED", and "All Public Schools" as "State". Variable
definitions are based on the \href{https://reportcard.msde.maryland.gov/Definitions/Index}{definitions from the MSDE website}.
}
\details{
Last updated 2024-02-22.
}
\keyword{datasets}
